@@ -1,21 +1,46 @@
-apply plugin: 'com.android.application'
-apply plugin: 'org.jetbrains.kotlin.android'
-apply plugin: 'kotlin-kapt'
-apply plugin: 'dagger.hilt.android.plugin'
+plugins {
+    id 'com.android.application'
+    id 'org.jetbrains.kotlin.android'
+    id 'kotlin-kapt'
+    id 'dagger.hilt.android.plugin'
+    id 'com.google.devtools.ksp' version '1.7.20-1.0.8'
+}
+
+def keystorePropertiesFile = rootProject.file("keystore.properties")
+def keystoreProperties = new Properties()
+keystoreProperties.load(new FileInputStream(keystorePropertiesFile))
 
 android {
-    compileSdkVersion 32
+    namespace 'com.abhishek.germanPocketDictionary'
+    compileSdkVersion 33
+
+    signingConfigs {
+        release {
+            storeFile file(keystoreProperties['storeFile'])
+            storePassword keystoreProperties['storePassword']
+            keyAlias keystoreProperties['keyAlias']
+            keyPassword keystoreProperties['keyPassword']
+        }
+    }
+
     defaultConfig {
         applicationId "com.abhishek.germanPocketDictionary"
-        minSdkVersion 21
+        minSdk 21
         targetSdkVersion 32
-        versionCode 13
-        versionName '1.1.3'
+        versionCode 2
+        versionName '2.0.0'
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 
         //Defining Log debugging
         buildConfigField "boolean", "LOG_DEBUG_MODE", "false"
         buildConfigField "boolean", "LOG_DEBUG_WITH_STACKTRACE_MODE", "false"
+
+        vectorDrawables {
+            useSupportLibrary true
+        }
+        signingConfig signingConfigs.release
+
+        archivesBaseName = "$applicationName-v$versionName"
     }
     buildTypes {
 
